+++ CHANGELOG.md
@@ -2,8 +2,9 @@ v.0.1.10
 - Multilingual models cache lifetime placed in Application\Mvc\Model::CACHE_LIFETIME constant;
 - Multilingual models cache data correctly clean after saving entity; 
 - Memcache host and port placed in configuration files;
+- Cms\Scanner for translations scanning only /app/modules/ and /app/views/ directories;
 - Minor fixes;
-- Added LICENSE.txt;
+- Added LICENSE.md;
 
 v.0.1.9
 - Semantic UI Bootstrap 3 theme integrated;